+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Full Changelog](https://github.com/folio-org/ui-courses/compare/v1.1.0...v1.0.0)
 
 * Upgrade to `@folio/stripes` v3.x, which is what FOLIO 2020q1 will use. Fixes UICR-58.
+* Remove non-functional reserve filters "Copyright tracking?" and "Status". Part of UICR-31.
 * Towards correct sorting for start-date, end-date, instructor and status: see UICR-49 and UICR-50, but note that these will not work until MODCR-21 and its underlying RMB issue are fixed.
 * Towards more correct searching in reserves, though underlying issues mean these changes will not yet help: see UICR-31.
 * Towards more correct sorting in reserves, though underlying issues mean  these changes will not yet help: see UICR-60.